move report form template into reports
