+++ src/posts/template_optimizer.py
@@ -328,45 +328,6 @@ def email_wrap(post: dict, val: str):
         return val
     return f'<a href="mailto:{ email }">{val}</a>'
 
-
-def generate_report_modal():
-    category_options = "\n".join(
-        f"""
-        <div>
-          <input type="radio" id="{category.name}" name="submitter_category" value="{category.value}" required>
-          <label for="{category.name}">{category.value}</label>
-        </div>
-        """ for category in SubmitterCategory
-    )
-    modal_html = f"""
-    <div id="modal_overlay" hidden>
-        <div id="report_modal" class="form" hidden>
-            <div class="modal_header">
-                <div class="modal_title">Report</div>
-                <div id="report_close" class="btn">Close</div>
-            </div>
-            <form class="form" id="report_form" action="" method="POST">
-                <div>
-                    <label for="submitter_category">Category:</label>
-                    {category_options}
-                </div>
-                <br>
-                <div>
-                    <label for="submitter_notes">Details:</label>
-                    <textarea id="submitter_notes" name="submitter_notes" cols="48" rows="8" maxlength="512" placeholder="Provide details about the issue."></textarea>
-                </div>
-                <br>
-                <div id="feedback_report"></div>
-                <input type="submit" value="Submit">
-            </form>
-        </div>
-    </div>
-    """
-    return modal_html
-
-report_modal_t = generate_report_modal()
-
-
 op_label = '<strong class="op_label">OP</strong>'
 def render_wrapped_post_t(wpt: dict): # wrapped_post_t
     is_op = wpt['op']
