+++ src/templates.py
@@ -2,7 +2,7 @@ from jinja2 import Environment, FileSystemLoader, select_autoescape
 from quart import get_flashed_messages, request, url_for
 
 from boards import board_objects
-from posts.template_optimizer import report_modal_t
+from moderation.report import report_modal_t
 from configs import app_conf, media_conf, site_conf, vanilla_search_conf, index_search_conf, tag_conf, archiveposting_conf, mod_conf, stats_conf
 from utils import make_src_path
 from utils.timestamps import ts_2_formatted
